+++ R/item_restscore.R
@@ -50,7 +50,7 @@
 #' @references
 #' Kreiner, S. (2011). A Note on Item–Restscore Association in Rasch Models. 
 #' *Applied Psychological Measurement, 35*(7), 557–561. 
-#' <https://doi.org/10.1177/0146621611410227>
+#' \doi{10.1177/0146621611410227}
 #'
 #' @export
 #'