+++ R/reliability.R
@@ -52,7 +52,7 @@ cronbach_alpha <- function(data) {
 #' @references
 #' Bignardi, G., Kievit, R., & Bürkner, P. C. (2025). A general method for
 #' estimating reliability using Bayesian Measurement Uncertainty. *PsyArXiv*.
-#' \doi{10.31234/osf.io/h54k8}
+#' \doi{10.31234/osf.io/h54k8_v1}
 #'
 #' @seealso [RMreliability()]
 #'
@@ -192,7 +192,7 @@ RMUreliability <- function(input_draws, level = 0.95, verbose = FALSE) {
 #' @references
 #' Bignardi, G., Kievit, R., & Bürkner, P. C. (2025). A general method for
 #' estimating reliability using Bayesian Measurement Uncertainty. *PsyArXiv*.
-#' \doi{10.31234/osf.io/h54k8}
+#' \doi{10.31234/osf.io/h54k8_v1}
 #'
 #' Mislevy, R. J. (1991). Randomization-Based Inference about Latent Variables
 #' from Complex Samples. *Psychometrika, 56*(2), 177-196.